Remove Assist command execution and chat feature
The changes remove tests and related code for command execution in the assist package and web command handling.
